Added analyzer
Necessary to collapse the specialized positional, keyword, and splat
classes into a single object. This will provide an enhanced -- and
consolidated -- Object API along with performance efficiencies in terms
of computation and memory usage.

Reek is being updated due to too many methods but worth the cost in
this case due to the above reasons.