Added builder
Necessary to build a parameter which is part of a method's greater
parameter signature. This will soon be used to build a method's
complete signature.

RuboCop cyclomatic complexity is being ignored because keeping all of
this logic together is more beneficial than tearing it apart further.

Reek's too many statements check is being ignored for the same reason
as stated above for RuboCop. Both are related.