Added defaulter
Necessary to calculate the default value for a parameter. This is only
necessary for optional positional or keyword parameters.

Convenience has been added to for nils, strings, symbols, and objects
as default values. The only caveat is that objects need to be prefixed
with "*" to allow them to pass though and not be confused as a string.
Objects must also be wrapped in strings to ensure lazy evaluation.

Reek's nil check is being ignored because if a nil *is* detected, it
must be converted to a string in order to show up, properly, as a
default value.