Added splat
Necessary to capture all possible arguments sent to a method for
splatting (i.e. single and double splats). The block is a slight
oddity in that it's not, technically, a splat but a forwarding of the
argument. Regardless, this is extremely useful when dynamically
building all arguments for dynamic messaging. Example:

Example:

    splat = Splat[
      positionals: [1, 2, 3],
      keywords: {a: 1, b: 2},
      proc { "test" }
    ]

    demo.public_send :call,
                     *splat.positionals,
                     **splat.keywords,
                     &splat.block