Fixed categorizer forwarded and bare splat argument handling
Simplfies the original logic where the supplied arguments must be:

1. An array.
2. A positional element within the array.

The above hasn't changed with this fix but the expected format of the
supplied arguments has been adjusted to account for bare splat,
forwarded, and named splat parameters. The following shows how the
arguments are now expected to be matched to the corresponding
parameters:

    def demo(*, **, &) = super
    # Parameters: [[:rest], [:keyrest], [:block, :&]]
    # Arguments:  [[], {}, &]

    def demo(...) = super
    # Parameters: [[:rest, :*], [:keyrest, :**], [:block, :&]]
    # Arguments:  [[], {}, &]

    def demo(*a, **b, &c) = super
    # Parameters: [[:rest, :a], [:keyrest, :b], [:block, :c]]
    # Arguments:  [[], {}, &]

These changes also account for `nil` arguments since one can -- and
will -- want to provide a `nil` argument to override a default value.

The spec descriptions were updated to focus on the arguments answered
back since that is what we care about most even though they are part
of the `Splat` struct as a delivery mechanism.