Refactored probe to use parameter array instead of hash
Necessary to fix a bug where multiple parameters of the same kind
would be overwritten when using a hash (since keys must be unique).
These changes focus on using the supplied parameters as the source of
information instead of mirroring the parameters with a secondary hash.

Support for hashes will be completely removed in a future commit.