Fixed categorizer nullable arguments
In an earlier fix (8e7b216b3a2e), `nil` arguments were being passed
through for situations in which someone might want to override an
optional parameter with a `nil` argument. The problem is that the
original fix assumed use of *implicit* `nil` arguments when not
provided but they must be *explicitly* provided to be used. A good
example of this is when you have the following method signature and
arguments:

    def demo(one, two = {}) = super

    demo 1          # Valid (i.e. `[1]`).
    demo 2, {b: 2}  # Valid (i.e. `[2, {b: 2}]`).
    demo 3, nil     # Valid but only if defined (i.e. `[3, nil]`).

This change also fixes a performance issue where if the arguments
array is empty, we immediately answer back an empty record instead of
iterating through all parameters which is an unnecessary operation.