Removed defaulter passthrough in favor of source parsing
Necessary to acquire the source code from a `Proc` since we don't need to evaluate but only extract the _literal_ contents for use when building method signatures later.

Procs, when used as default values, are not meant to be defined with parameters because they'll never be called. This is also why lambdas are not allowed since there is no need to check for arguments. To prevent parameter use, this checks for arity and immediately fails if greater than zero so folks can correct behavior early and quickly.

This also eliminates the need to use the `passthrough` which was awkward and unsafe to begin with. Now procs can serve as a delivery mechanism without having folks learn new syntax. This is made possible by the recently added `Sourcer` class.

The signature specs were updated accordingly to account for these changes.

Milestone: major