Removed defaulter passthrough in favor of syntax parsing
Necessary to improve security by using the VM to acquire the source code from the proc/lambda since we don't need to evaluate but only extract the _literal_ contents for use when building method signatures later.

The source node is always `2` for procs/lambdas. This is unfortunate since `2` ends up being a magic number but, for this use case, I opted to stick with the magic number since the second node is always the source (body) of the proc/lambda.

Procs/lambdas, when used as default values, are not meant to be defined with parameters because they'll never be called. To prevent parameter use, this will also check for procs and immediately fail so folks can correct behavior early and quickly. The only way a `Proc` will appear as a value is if the `Signature` finds one with a non-zero arity.

This also eliminates the need to use the `passthrough` which was awkward and unsafe to begin with. Now procs can serve as a delivery mechanism without having folks learn new syntax.

Milestone: major