Removed lesser used probe methods
The splat and block methods are no longer used and are safe to delete.

For keyword slicing, the method is no longer needed because this has been replaced by the `+#keywords_for+` method.

For hash conversion, this is a revert of 0e360d99f889 (Added probe hash conversion) where hash conversion is too confusing because you can have dupicalte types which a hash won't respect while a double nested array will.

Milestone: major