Fixed builder to account for no keywords
Necessary to support the no keywords parameter despite being rare in use. This also makes the name optional since supplying a name for no keywords isn't supported by the Ruby parser.

The specs were updated to use the same parameter name in order to be consistent with terminology used in the rest of the test suite.

RuboCop was updated to ignore method length because splitting up the case statement would degrade readability rather than improve it.

Milestone: patch