Added signature inheritor
Necessary to ensure parameters are respectful of the super method's signature by merging the super and sub method parameters together.

RuboCop and Reek were updated to exclude the issues with the `merge` method. This is definitely a concern but is excluded, in this case, because keeping this logic self contained is better than splitting out further (at least for now, this might change in the future).

Milestone: minor