Added super signature
Necessary for building the arguments for forwarding via the `super` keyword. This is meant to be used in conjunction with the sub-method signature construction when working with inheritance.

RuboCop and Reek were updated to exclude the issues with the `merge` method. This is definitely a concern but is excluded, in this case, because keeping this logic self contained is better than splitting out further (at least for now, this might change in the future).

Milestone: minor