Added signature sourcer
Necessary to acquire the source code of a `Proc`. This will soon be used so default values for parameters can be wrapped in a `Proc` for lazy evaluation of complex objects.

There are situations -- like when using IRB, `eval`, C code, and so forth -- where the source path isn't a path and will answer a string wrapped in parenthesis instead. This accounts for those situations by ensuring the path starts with a file separator and answer `nil` if otherwise so the generated signature remains syntactically correct.

Milestone: minor